Updated Classes

+++ HOME/Results.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 public class Results
 {
     public int Uload { get; set; }
-    public int UPoints { get; set; }
+    public double UPoints { get; set; }
     public int UShareable { get; set; }
     public int UMagicData { get; set; }
     public int UEverydayGB { get; set; }
